requirements.txt
Django==5.2
djangorestframework==3.15.2
python-dotenv==1.1.0
drf_spectacular==0.28.0
Pillow==11.2.1
psycopg2-binary==2.9.10
djangorestframework_simplejwt==5.5.0
requests==2.32.3
django-cors-headers==4.3.1
django-filter>=23.3
django-dbml==0.10.1